Feature/update practice issues

See merge request g-tungnguyen/react-training!6

+++ booking-management-system/src/components/SideBar/RoomCategory/RoomCategory.tsx
@@ -35,8 +35,9 @@ export const RoomCategory = () => {
   //Delete room category
   const deleteCategory = async (id: string) => {
     if (window.confirm('You sure to delete?')) {
-      dispatch(deleteRoom({ id }));
-      await api.delete(`/roomCategory/${id}`);
+      if (await api.delete(`/roomCategory/${id}`)) {
+        dispatch(deleteRoom({ id }));
+      }
     }
   };
